Migrate all prefab inline functions to their own registry per category
